refactor: move device management into `DeviceService`
